refactor: movendo arquivo pra pacote novo
